+++ client/fight/interface/MapVisualiserPlayerWithFight.h
@@ -10,9 +10,9 @@ class MapVisualiserPlayerWithFight : public MapVisualiserPlayer
     Q_OBJECT
 public:
     explicit MapVisualiserPlayerWithFight(const bool &centerOnPlayer=true, const bool &debugTags=false, const bool &useCache=true, const bool &OpenGL=false);
-    void setBotsAlreadyBeaten(const QSet<quint32> &botAlreadyBeaten);
-    void addBeatenBotFight(const quint32 &botFightId);
-    bool haveBeatBot(const quint32 &botFightId) const;
+    void setBotsAlreadyBeaten(const QSet<quint16> &botAlreadyBeaten);
+    void addBeatenBotFight(const quint16 &botFightId);
+    bool haveBeatBot(const quint16 &botFightId) const;
     void addRepelStep(const quint32 &repel_step);
 protected slots:
     virtual void keyPressParse();
@@ -20,7 +20,7 @@ protected slots:
     virtual bool canGoTo(const CatchChallenger::Direction &direction,CatchChallenger::CommonMap map,COORD_TYPE x,COORD_TYPE y,const bool &checkCollision);
     virtual void resetAll();
 protected:
-    QSet<quint32> botAlreadyBeaten;
+    QSet<quint16> botAlreadyBeaten;
     quint32 repel_step;
 signals:
     void repelEffectIsOver() const;
